Move tasks to individual cogs files
